move file_path_manager to server_common
