+++ BlockServer/core/config_holder.py
@@ -25,7 +25,7 @@ from BlockServer.config.configuration import Configuration
 from BlockServer.config.group import Group
 from BlockServer.config.metadata import MetaData
 from BlockServer.core.constants import DEFAULT_COMPONENT, GRP_NONE
-from BlockServer.core.file_path_manager import FILEPATH_MANAGER
+from server_common.file_path_manager import FILEPATH_MANAGER
 from BlockServer.core.macros import PVPREFIX_MACRO
 from BlockServer.fileIO.file_manager import ConfigurationFileManager
 from server_common.utilities import print_and_log
